refactor: finalize analytics migration
