+++ src/stable_yield_lab/analytics/__init__.py
@@ -1,0 +1,11 @@
+"""Analytics subpackage bundling portfolio, performance, and risk helpers."""
+
+from . import attribution, metrics, performance, portfolio, risk
+
+__all__ = [
+    "attribution",
+    "metrics",
+    "performance",
+    "portfolio",
+    "risk",
+]
